+++ openeo_driver/processes.py
@@ -7,7 +7,6 @@ import inspect
 from pathlib import Path
 from typing import Any, Callable, Collection, Dict, List, Optional, Tuple, Union, Iterable
 
-from openeo_driver.datacube import DriverDataCube
 from openeo_driver.errors import (
     OpenEOApiException,
     ProcessParameterInvalidException,
@@ -382,6 +381,36 @@ ArgumentValue = Any
 Validator = Callable[[Any], bool]
 
 
+@functools.lru_cache
+def _describe_type(type_: Union[type, Tuple[type, ...]]) -> str:
+    """
+    Map Python type (or tuple of types) to more user-friendly,
+    and less Python-specific name or description,
+    to be used in `ProcessParameterInvalid` errors.
+    """
+
+    # Local import to avoid import cycles
+    from openeo_driver.datacube import DriverDataCube
+    from openeo_driver.datacube import DriverVectorCube
+
+    type_map = {
+        int: "integer",
+        float: "float",
+        str: "string",
+        dict: "dictionary/mapping",
+        type(None): "null/None",
+        DriverDataCube: "raster cube",
+        DriverVectorCube: "vector cube",
+    }
+
+    if type_ in type_map:
+        return type_map[type_]
+    elif isinstance(type_, tuple):
+        return " or ".join(_describe_type(t) for t in type_)
+    else:
+        return str(type_)
+
+
 class ProcessArgs(dict):
     """
     Wrapper for process argument extraction with proper exception throwing.
@@ -392,6 +421,7 @@ class ProcessArgs(dict):
 
     def __init__(self, args: dict, process_id: Optional[str] = None):
         super().__init__(args)
+        # Current process id
         self.process_id = process_id
 
     @classmethod
@@ -410,19 +440,31 @@ class ProcessArgs(dict):
         name: str,
         *,
         expected_type: Optional[Union[type, Tuple[type, ...]]] = None,
+        expected_type_name: Optional[str] = None,
         validator: Optional[Validator] = None,
     ) -> ArgumentValue:
         """
         Get a required argument by name.
 
         Originally: `extract_arg`.
+
+        :param name: argument name
+        :param default: default value or a function/factory to generate the default value
+        :param expected_type: expected class (or list of multiple options) the value should be (unless it's None)
+        :param expected_type_name: user friendly description of expected type(s)
+        :param validator: optional validation callable
         """
-        # TODO: add option for type check too
         try:
             value = self[name]
         except KeyError:
             raise ProcessParameterRequiredException(process=self.process_id, parameter=name) from None
-        self._check_value(name=name, value=value, expected_type=expected_type, validator=validator)
+        self._check_value(
+            name=name,
+            value=value,
+            expected_type=expected_type,
+            expected_type_name=expected_type_name,
+            validator=validator,
+        )
         return value
 
     def _check_value(
@@ -431,14 +473,17 @@ class ProcessArgs(dict):
         name: str,
         value: Any,
         expected_type: Optional[Union[type, Tuple[type, ...]]] = None,
+        expected_type_name: Optional[str] = None,
         validator: Optional[Validator] = None,
     ):
         if expected_type:
             if not isinstance(value, expected_type):
-                if expected_type is DriverDataCube:
-                    expected_type = "raster cube"
+                expected_type_name = expected_type_name or _describe_type(expected_type)
+                actual_type_name = _describe_type(type(value))
                 raise ProcessParameterInvalidException(
-                    parameter=name, process=self.process_id, reason=f"Expected {expected_type} but got {type(value)}."
+                    parameter=name,
+                    process=self.process_id,
+                    reason=f"Expected {expected_type_name} but got {actual_type_name}.",
                 )
         if validator:
             reason = None
@@ -461,6 +506,7 @@ class ProcessArgs(dict):
         default: Union[Any, Callable[[], Any]] = None,
         *,
         expected_type: Optional[Union[type, Tuple[type, ...]]] = None,
+        expected_type_name: Optional[str] = None,
         validator: Optional[Validator] = None,
     ) -> ArgumentValue:
         """
@@ -469,6 +515,7 @@ class ProcessArgs(dict):
         :param name: argument name
         :param default: default value or a function/factory to generate the default value
         :param expected_type: expected class (or list of multiple options) the value should be (unless it's None)
+        :param expected_type_name: user friendly description of expected type(s)
         :param validator: optional validation callable
         """
         if name in self:
@@ -476,7 +523,13 @@ class ProcessArgs(dict):
         else:
             value = default() if callable(default) else default
         if value is not None:
-            self._check_value(name=name, value=value, expected_type=expected_type, validator=validator)
+            self._check_value(
+                name=name,
+                value=value,
+                expected_type=expected_type,
+                expected_type_name=expected_type_name,
+                validator=validator,
+            )
 
         return value