+++ tests/test_views_execute.py
@@ -1790,7 +1790,7 @@ def test_run_udf_on_aggregate_spatial(api, udf_code):
             CURRENT_PY3x,
             (
                 "ProcessParameterInvalid",
-                "The value passed for parameter 'runtime' in process 'run_udf' is invalid: Expected <class 'str'> but got <class 'NoneType'>.",
+                "The value passed for parameter 'runtime' in process 'run_udf' is invalid: Expected string but got null/None.",
             ),
         ),
     ],